+++ dbversioning/versionedDb.py
@@ -1,12 +1,12 @@
 import hashlib
 import os
 from shutil import rmtree
-from typing import List
+from typing import List, Optional
 
 from dbversioning.errorUtil import (
     VersionedDbExceptionRepoVersionNumber,
     VersionedDbExceptionRepoDoesNotExits,
-    VersionedDbExceptionRepoNameInvalid, VersionedDbExceptionSqlNamingError)
+    VersionedDbExceptionRepoNameInvalid, VersionedDbExceptionSqlNamingError, VersionedDbExceptionInvalidSqlName)
 from dbversioning.osUtil import dir_exists
 from dbversioning.repositoryconf import (
     DATA_DUMP_DIR, ROLLBACK_FILE_ENDING)
@@ -40,7 +40,7 @@ class SchemaSnapshotVersion(object):
         self.minor = None
         self.maintenance = None
 
-        _set_version_info(os.path.basename(self._version_path), self)
+        self._set_version_info(os.path.basename(self._version_path))
         file_array = os.path.splitext(os.path.basename(ss_version_path))
 
         self.name = file_array[0]
@@ -56,6 +56,24 @@ class SchemaSnapshotVersion(object):
     @property
     def sql_file(self):
         return self._version_path
+
+    def _set_version_info(self, file_name):
+        ver_array = file_name.split(".")
+        file_ext = ver_array.pop()
+
+        if len(ver_array) < 1 or file_ext != "sql":
+            raise VersionedDbExceptionInvalidSqlName(file_name)
+
+        self.major = _try_int_returning(ver_array[0])
+
+        if self.major is not None and len(ver_array) > 2:
+            self.minor = _try_int_returning(ver_array[1])
+            if self.minor is not None and len(ver_array) > 3:
+                self.maintenance = _try_int_returning(ver_array[2])
+                self.name = ".".join(ver_array[-3:])
+
+        if self.major is None:
+            self.name = ".".join(ver_array)
 
 
 class Version(object):
@@ -216,3 +234,12 @@ def get_file_hash(file_path) -> str:
             sha1.update(data)
 
     return sha1.hexdigest()
+
+
+def _try_int_returning(val) -> Optional[int]:
+    try:
+        rtn = int(val)
+    except ValueError:
+        rtn = None
+
+    return rtn